+++ GriegSwing/src/pl/edu/agh/ki/grieg/widgets/swing/SwingCanvas.java
@@ -24,7 +24,7 @@ public class SwingCanvas extends AbstractCanvas {
         // empty
     }
 
-    public SwingCanvas(float width, float height) {
+    public SwingCanvas(double width, double height) {
         super(width, height);
     }
